# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AddLibraryAddonsRequest": {
        "additionalProperties": false,
        "properties": {
          "proposal": {
            "$ref": "#/components/schemas/ProposalDocument"
          },
          "services": {
            "description": "Service names from the add-on library.",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Services",
            "type": "array"
          }
        },
        "required": [
          "proposal",
          "services"
        ],
        "title": "AddLibraryAddonsRequest",
        "type": "object"
      },
      "AddLibraryItemsRequest": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "description": "Item names from the scope library.",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "proposal": {
            "$ref": "#/components/schemas/ProposalDocument"
          },
          "status": {
            "default": "In Scope",
            "enum": [
              "In Scope",
              "Out of Scope",
              "Optional Add-on",
              "Client Responsibility"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "proposal",
          "items"
        ],
        "title": "AddLibraryItemsRequest",
        "type": "object"
      },
...
      "AddonRow": {
        "additionalProperties": false,
        "properties": {
          "add_to_total": {
            "default": false,
            "title": "Add to Total",
            "type": "boolean"
          },
          "billing": {
            "default": "One Time",
            "enum": [
              "One Time",
              "Monthly",
              "Per Unit",
              "Per Year",
              "Hourly"
            ],
            "title": "Billing",
            "type": "string"
          },
          "cost": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Cost",
            "type": "number"
          },
          "includes": {
            "default": "",
            "title": "Includes",
            "type": "string"
          },
          "service": {
            "default": "",
            "title": "Service",
            "type": "string"
          }
        },
        "title": "AddonRow",
        "type": "object"
      },
...
      "AssumptionRow": {
        "additionalProperties": false,
        "properties": {
          "assumption": {
            "default": "",
            "title": "Assumption",
            "type": "string"
          }
        },
        "title": "AssumptionRow",
        "type": "object"
      },
...
      "CoverageRow": {
        "additionalProperties": false,
        "properties": {
          "coverage": {
            "enum": [
              "Covered",
              "Not covered",
              "Available as add-on",
              "Client responsibility",
              "Excluded"
            ],
            "title": "Coverage",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "requirement": {
            "title": "Requirement",
            "type": "string"
          },
          "scope_items": {
            "description": "Comma-separated IDs of the scope items that cite this requirement.",
            "title": "Scope Items",
            "type": "string"
          }
        },
        "required": [
          "id",
          "requirement",
          "scope_items",
          "coverage"
        ],
        "title": "CoverageRow",
        "type": "object"
      },
...
      "GlossaryRow": {
        "additionalProperties": false,
        "properties": {
          "definition": {
            "default": "",
            "title": "Definition",
            "type": "string"
          },
          "term": {
            "default": "",
            "title": "Term",
            "type": "string"
          }
        },
        "title": "GlossaryRow",
        "type": "object"
      },
      "GovernanceRow": {
        "additionalProperties": false,
        "properties": {
          "cadence": {
            "default": "",
            "title": "Cadence",
            "type": "string"
          },
          "meeting": {
            "default": "",
            "title": "Meeting / Report",
            "type": "string"
          },
          "participants": {
            "default": "",
            "title": "Participants",
            "type": "string"
          },
          "purpose": {
            "default": "",
            "title": "Purpose",
            "type": "string"
          }
        },
        "title": "GovernanceRow",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JepImportRequest": {
        "additionalProperties": false,
        "properties": {
          "jep": {
            "additionalProperties": {
              "items": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              "type": "array"
            },
            "description": "Joint Evaluation Plan sections, as exported by that module.",
            "title": "Jep",
            "type": "object"
          },
          "proposal": {
            "$ref": "#/components/schemas/ProposalDocument"
          },
          "prospect_name": {
            "default": "",
            "title": "Prospect Name",
            "type": "string"
          }
        },
        "required": [
          "proposal",
          "jep"
        ],
        "title": "JepImportRequest",
        "type": "object"
      },
      "KpiRow": {
        "additionalProperties": false,
        "properties": {
          "definition": {
            "default": "",
            "title": "Definition",
            "type": "string"
          },
          "kpi": {
            "default": "",
            "title": "KPI",
            "type": "string"
          },
          "target": {
            "default": "",
            "title": "Target",
            "type": "string"
          },
          "type": {
            "default": "Primary",
            "enum": [
              "Primary",
              "Secondary"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "title": "KpiRow",
        "type": "object"
      },
...
      "MethodologyRow": {
        "additionalProperties": false,
        "properties": {
          "activities": {
            "default": "",
            "title": "Activities",
            "type": "string"
          },
          "outputs": {
            "default": "",
            "title": "Outputs",
            "type": "string"
          },
          "phase": {
            "default": "",
            "title": "Phase",
            "type": "string"
          }
        },
        "title": "MethodologyRow",
        "type": "object"
      },
      "MilestoneRow": {
        "additionalProperties": false,
        "properties": {
          "deliverables": {
            "default": "",
            "title": "Deliverables",
            "type": "string"
          },
          "due_on_or_before": {
            "default": "",
            "title": "Due On or Before",
            "type": "string"
          },
          "fee": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Fee",
            "type": "number"
          },
          "milestone": {
            "default": "",
            "title": "Milestone",
            "type": "string"
          },
          "start_on_or_before": {
            "default": "",
            "title": "Start On or Before",
            "type": "string"
          }
        },
        "title": "MilestoneRow",
        "type": "object"
      },
...
      "ObjectiveRow": {
        "additionalProperties": false,
        "properties": {
          "objective": {
            "default": "",
            "title": "Objective",
            "type": "string"
          },
          "success_measure": {
            "default": "",
            "title": "Success Measure",
            "type": "string"
          }
        },
        "title": "ObjectiveRow",
        "type": "object"
      },
...
      "Party": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "default": "",
            "title": "Address",
            "type": "string"
          },
          "company": {
            "default": "",
            "title": "Company",
            "type": "string"
          },
          "contact": {
            "default": "",
            "title": "Contact",
            "type": "string"
          },
          "email": {
            "default": "",
            "title": "Email",
            "type": "string"
          },
          "phone": {
            "default": "",
            "title": "Phone",
            "type": "string"
          }
        },
        "title": "Party",
        "type": "object"
      },
...
      "PricingSummary": {
        "additionalProperties": false,
        "properties": {
          "discount": {
            "title": "Discount",
            "type": "number"
          },
          "discount_pct": {
            "title": "Discount Pct",
            "type": "number"
          },
          "milestone_total": {
            "title": "Milestone Total",
            "type": "number"
          },
          "recurring": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Recurring add-on totals keyed by billing period.",
            "title": "Recurring",
            "type": "object"
          },
          "selected_addons_one_time": {
            "title": "Selected Addons One Time",
            "type": "number"
          },
          "subtotal": {
            "title": "Subtotal",
            "type": "number"
          },
          "total": {
            "title": "Total",
            "type": "number"
          }
        },
        "required": [
          "milestone_total",
          "selected_addons_one_time",
          "subtotal",
          "discount_pct",
          "discount",
          "total",
          "recurring"
        ],
        "title": "PricingSummary",
        "type": "object"
      },
...
      "ProposalAnalysis": {
        "additionalProperties": false,
        "description": "Everything the builder shows alongside the document, in one round trip.",
        "properties": {
          "coverage": {
            "items": {
              "$ref": "#/components/schemas/CoverageRow"
            },
            "title": "Coverage",
            "type": "array"
          },
          "pricing": {
            "$ref": "#/components/schemas/PricingSummary"
          },
          "scope_by_status": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ScopeRow"
              },
              "type": "array"
            },
            "propertyNames": {
              "enum": [
                "In Scope",
                "Out of Scope",
                "Optional Add-on",
                "Client Responsibility"
              ]
            },
            "title": "Scope By Status",
            "type": "object"
          },
          "warnings": {
            "description": "Gaps a client would ask about. Advisory; they never block generation.",
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "pricing",
          "coverage",
          "scope_by_status",
          "warnings"
        ],
        "title": "ProposalAnalysis",
        "type": "object"
      },
      "ProposalDocument": {
        "additionalProperties": false,
        "description": "A complete proposal / SOW. Start from ``GET /proposals/default``.",
        "properties": {
          "acceptance_process": {
            "default": "",
            "title": "Acceptance Process",
            "type": "string"
          },
          "addons": {
            "items": {
              "$ref": "#/components/schemas/AddonRow"
            },
            "title": "Addons",
            "type": "array"
          },
          "assumptions": {
            "items": {
              "$ref": "#/components/schemas/AssumptionRow"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "change_process": {
            "default": "",
            "title": "Change Process",
            "type": "string"
          },
          "client": {
            "$ref": "#/components/schemas/Party"
          },
          "company_overview": {
            "default": "",
            "title": "Company Overview",
            "type": "string"
          },
          "confidentiality": {
            "default": "",
            "title": "Confidentiality",
            "type": "string"
          },
          "currency": {
            "default": "USD",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "current_state": {
            "default": "",
            "title": "Current State",
            "type": "string"
          },
          "discount_pct": {
            "default": 0.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Discount Pct",
            "type": "number"
          },
          "document_type": {
            "default": "Proposal + Statement of Work",
            "enum": [
              "Proposal + Statement of Work",
              "Proposal",
              "Statement of Work"
            ],
            "title": "Document Type",
            "type": "string"
          },
          "effective_date": {
            "default": "",
            "title": "Effective Date",
            "type": "string"
          },
          "executive_summary": {
            "default": "",
            "title": "Executive Summary",
            "type": "string"
          },
          "expenses": {
            "default": "",
            "title": "Expenses",
            "type": "string"
          },
          "expiration_date": {
            "default": "",
            "title": "Expiration Date",
            "type": "string"
          },
          "glossary": {
            "items": {
              "$ref": "#/components/schemas/GlossaryRow"
            },
            "title": "Glossary",
            "type": "array"
          },
          "governance": {
            "items": {
              "$ref": "#/components/schemas/GovernanceRow"
            },
            "title": "Governance",
            "type": "array"
          },
          "issue_date": {
            "default": "",
            "title": "Issue Date",
            "type": "string"
          },
          "kpis": {
            "items": {
              "$ref": "#/components/schemas/KpiRow"
            },
            "title": "Kpis",
            "type": "array"
          },
          "methodology": {
            "items": {
              "$ref": "#/components/schemas/MethodologyRow"
            },
            "title": "Methodology",
            "type": "array"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneRow"
            },
            "title": "Milestones",
            "type": "array"
          },
          "objectives": {
            "items": {
              "$ref": "#/components/schemas/ObjectiveRow"
            },
            "title": "Objectives",
            "type": "array"
          },
          "payment_terms": {
            "default": "",
            "title": "Payment Terms",
            "type": "string"
          },
          "problem_statement": {
            "default": "",
            "title": "Problem Statement",
            "type": "string"
          },
          "rate_card": {
            "items": {
              "$ref": "#/components/schemas/RateCardRow"
            },
            "title": "Rate Card",
            "type": "array"
          },
          "reference_number": {
            "default": "",
            "title": "Reference Number",
            "type": "string"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementRow"
            },
            "title": "Requirements",
            "type": "array"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskRow"
            },
            "title": "Risks",
            "type": "array"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "scope": {
            "items": {
              "$ref": "#/components/schemas/ScopeRow"
            },
            "title": "Scope",
            "type": "array"
          },
          "sections": {
            "additionalProperties": {
              "type": "boolean"
            },
            "description": "Optional sections to include: executive_summary, company_overview, glossary, kpis, risks, addons, rate_card, signatures.",
            "title": "Sections",
            "type": "object"
          },
          "seller": {
            "$ref": "#/components/schemas/Party"
          },
          "solution_overview": {
            "default": "",
            "title": "Solution Overview",
            "type": "string"
          },
          "team": {
            "items": {
              "$ref": "#/components/schemas/TeamRow"
            },
            "title": "Team",
            "type": "array"
          },
          "techniques": {
            "items": {
              "$ref": "#/components/schemas/TechniqueRow"
            },
            "title": "Techniques",
            "type": "array"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineRow"
            },
            "title": "Timeline",
            "type": "array"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolRow"
            },
            "title": "Tools",
            "type": "array"
          },
          "valid_until": {
            "default": "",
            "title": "Valid Until",
            "type": "string"
          }
        },
        "title": "ProposalDocument",
        "type": "object"
      },
...
      "ProposalRequest": {
        "additionalProperties": false,
        "properties": {
          "proposal": {
            "$ref": "#/components/schemas/ProposalDocument"
          }
        },
        "required": [
          "proposal"
        ],
        "title": "ProposalRequest",
        "type": "object"
      },
...
      "ScopeRow": {
        "additionalProperties": false,
        "properties": {
          "acceptance_criteria": {
            "default": "",
            "title": "Acceptance Criteria",
            "type": "string"
          },
          "id": {
            "default": "",
            "description": "Assigned (S-01, S-02, ...) when blank.",
            "title": "ID",
            "type": "string"
          },
          "item": {
            "default": "",
            "title": "Item",
            "type": "string"
          },
          "milestone": {
            "default": "",
            "title": "Milestone",
            "type": "string"
          },
          "quantity_limit": {
            "default": "",
            "title": "Quantity / Limit",
            "type": "string"
          },
          "requirement_ids": {
            "default": "",
            "description": "Comma-separated requirement IDs this item covers.",
            "title": "Req. ID",
            "type": "string"
          },
          "status": {
            "default": "In Scope",
            "enum": [
              "In Scope",
              "Out of Scope",
              "Optional Add-on",
              "Client Responsibility"
            ],
            "title": "Status",
            "type": "string"
          },
          "tools_techniques": {
            "default": "",
            "title": "Tools / Techniques",
            "type": "string"
          },
          "workstream": {
            "default": "",
            "title": "Workstream",
            "type": "string"
          }
        },
        "title": "ScopeRow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WrdImportRequest": {
        "additionalProperties": false,
        "properties": {
          "proposal": {
            "$ref": "#/components/schemas/ProposalDocument"
          },
          "wrd": {
            "additionalProperties": {
              "items": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              "type": "array"
            },
            "description": "Requirements Document sections, as exported by that module.",
            "title": "Wrd",
            "type": "object"
          }
        },
        "required": [
          "proposal",
          "wrd"
        ],
        "title": "WrdImportRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HubSpotToken": {
        "description": "HubSpot private app access token (read-only scopes). Never stored or logged.",
        "in": "header",
        "name": "X-HubSpot-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "summary": "HubSpot implementation toolkit for solutions architects. Read-only against HubSpot.",
    "title": "ArchitectureScope API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/proposals/analysis": {
      "post": {
        "operationId": "proposals_analyse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalAnalysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pricing, requirement coverage, scope groups and warnings",
        "tags": [
          "proposals"
        ]
      }
    },
...
    "/api/v1/proposals/import/jep": {
      "post": {
        "operationId": "proposals_import_jep",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JepImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge a Joint Evaluation Plan into the proposal",
        "tags": [
          "proposals"
        ]
      }
    },
    "/api/v1/proposals/import/wrd": {
      "post": {
        "operationId": "proposals_import_wrd",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrdImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge a Requirements Document into the proposal",
        "tags": [
          "proposals"
        ]
      }
    },
...
    "/api/v1/proposals/pdf": {
      "post": {
        "operationId": "proposals_render_pdf",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render the branded proposal / SOW PDF",
        "tags": [
          "proposals"
        ]
      }
    },
    "/api/v1/proposals/scope/from-requirements": {
      "post": {
        "operationId": "proposals_draft_scope",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft scope items for uncovered requirements",
        "tags": [
          "proposals"
        ]
      }
    },
    "/api/v1/proposals/scope/library-addons": {
      "post": {
        "operationId": "proposals_add_library_addons",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLibraryAddonsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add add-on services (as Optional scope items)",
        "tags": [
          "proposals"
        ]
      }
    },
    "/api/v1/proposals/scope/library-items": {
      "post": {
        "operationId": "proposals_add_library_items",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLibraryItemsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add scope-library items",
        "tags": [
          "proposals"
        ]
      }
    },
    "/api/v1/proposals/template": {
      "post": {
        "operationId": "proposals_as_template",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Strip deal-specific content, keeping reusable boilerplate",
        "tags": [
          "proposals"
        ]
      }
    },
...
  }
}